handover fixed and test api
