+++ backend/app/app/api/api_v1/endpoints/__init__.py
@@ -9,3 +9,4 @@ from .UE import router
 from .qosInformation import router
 from .qosMonitoring import router
 from .monitoringevent import router
+from .tests import router
