+++ backend/app/app/ui/sidebar.html
@@ -47,6 +47,9 @@
             <li class="nav-item"><a class="nav-link" href="/docs" target="_blank">
               NEF_Emulator</a>
             </li>
+            <li class="nav-item"><a class="nav-link" href="/test/docs" target="_blank">
+              Test APIs</a>
+            </li>
           </ul>
         </li>
         <li class="nav-group" aria-expanded="false"><a class="nav-link nav-group-toggle" href="#">
